crawl4ai-service/requirements.txt
fastapi
uvicorn[standard]
crawl4ai
python-dotenv
google-generativeai
